update json library
